Client - Added Firebase cache folder to .gitignore - Fixed reloadProduct & added reloadCarts to get emits from Server using Socket.io - Added currency formater and coverter from USD to IDR. - Added addtional CSS class of CartContainer. - Added conditions for Dashboard component, such as wait for user data to be fetched. - Fixed See More button on ProductContainer component in order to be inline. - Addded currency formatter on ProductDetail component. - Created QtyForm to display and maintain cart's amounts, etc. - Fixed checkLoginStatus method.
Server
- Fixed CRUD Cart model.
- Fixed User Check controller to include Cart model and return User data excluding password.
- Changed Product price data type from INTEGER to DECIMAL(12, 2).
- Added Amount property on Cart's migration & model file.
- Edited seeders to use USD instead of IDR.
- Succcessfully implemented Cart edit and delete.

+++ server/models/product.js
@@ -17,7 +17,7 @@ module.exports = (sequelize, DataTypes) => {
         type: DataTypes.STRING
       },
       price: {
-        type: DataTypes.INTEGER
+        type: DataTypes.DECIMAL(12, 2)
       },
       stock: {
         type: DataTypes.INTEGER
